consolidate graphql stuff
