+++ package.json
@@ -21,6 +21,7 @@
     "graphql": "^15.0.0",
     "graphql-request": "^1.8.2",
     "graphql-tag": "^2.10.3",
+    "graphql-tools": "^5.0.0",
     "next": "9.4.1",
     "react": "^16.8.0",
     "react-dom": "^16.8.0",
@@ -32,10 +33,10 @@
     "swr": "^0.2.0"
   },
   "devDependencies": {
-    "@graphql-codegen/cli": "1.13.5",
-    "@graphql-codegen/typescript": "1.13.5",
+    "@graphql-codegen/cli": "1.14.0",
+    "@graphql-codegen/typescript": "1.14.0",
     "@graphql-codegen/typescript-operations": "^1.13.5",
-    "@graphql-codegen/typescript-resolvers": "1.13.5",
+    "@graphql-codegen/typescript-resolvers": "1.14.0",
     "@types/node": "^14.0.1",
     "@types/react": "^16.9.35",
     "@types/react-dom": "^16.9.8",
