+++ src/graphql/index.ts
@@ -1,4 +1,3 @@
-export * from './generated';
 export type { PantryContext } from './pantry.graphql';
 export { schema as typeDefs } from './pantry.graphql';
-export { GetPantryQueryString } from './queries/get-pantry';
+export { schema } from './schema';
