+++ src/graphql/resolvers/index.ts
@@ -1,0 +1,10 @@
+import { Resolvers } from '../__generated__';
+import { getPantry } from '../../db';
+
+export const resolvers: Resolvers = {
+  Query: {
+    pantry: (_parent, _args, _context) => {
+      return getPantry();
+    },
+  },
+};
