+++ src/graphql/schema.ts
@@ -1,0 +1,8 @@
+import { makeExecutableSchema } from 'graphql-tools';
+import { schema as typeDefs } from './pantry.graphql';
+import { resolvers } from './resolvers';
+
+export const schema = makeExecutableSchema({
+  typeDefs,
+  resolvers,
+});
