+++ src/pages/index.tsx
@@ -15,10 +15,9 @@ import {
 } from '../components';
 import {
   GetPantryQueryResult,
-  GetPantryQueryString,
   GetPantryQueryVariables,
-} from '../schema';
-import { FOOD_IMAGES } from '../lib/foodImages';
+} from '../graphql/__generated__';
+import { GetPantryQueryString } from '../graphql/queries/get-pantry';
 
 async function pantryLoader(query: string, variables: GetPantryQueryVariables) {
   return request<GetPantryQueryResult>('/api/graphql', query, variables);
@@ -85,21 +84,73 @@ export default function Pantry(props: Props) {
 export const getServerSideProps: GetServerSideProps = async () => {
   const { getPantry } = await import('../db');
   const pantryItems = await getPantry();
-  const foodImagesMap = FOOD_IMAGES.map((image) => {
-    const foodName = image.replace('icons8-', '').replace('-100.png', '');
-    return {
-      image,
-      name: foodName.replace('-', ''),
-    };
-  });
+  const foodImages = getFoodImages();
+
   const initialProps: Props = {
     initialData: {
       pantry: pantryItems,
     },
-    foodImages: foodImagesMap,
+    foodImages,
   };
 
   return {
     props: initialProps,
   };
 };
+
+function getFoodImages() {
+  return [
+    'icons8-asparagus-100.png',
+    'icons8-avocado-100.png',
+    'icons8-bacon-100.png',
+    'icons8-big-carrot-100.png',
+    'icons8-bread-100.png',
+    'icons8-cauliflower-100.png',
+    'icons8-cheese-100.png',
+    'icons8-chili-pepper-100.png',
+    'icons8-chocolate-bar-100.png',
+    'icons8-cinnamon-sticks-100.png',
+    'icons8-corn-100.png',
+    'icons8-cucumber-100.png',
+    'icons8-dressing-100.png',
+    'icons8-egg-carton-100.png',
+    'icons8-fish-skeleton-100.png',
+    'icons8-flour-100-2.png',
+    'icons8-flour-100.png',
+    'icons8-garlic-100.png',
+    'icons8-grape-100.png',
+    'icons8-greenery-100.png',
+    'icons8-honey-100.png',
+    'icons8-ice-cream-in-waffle-cone-100.png',
+    'icons8-jam-100.png',
+    'icons8-lemon-100.png',
+    'icons8-lettuce-100.png',
+    'icons8-meat-100.png',
+    'icons8-meat-sausage-100.png',
+    'icons8-milk-bottle-100.png',
+    'icons8-mushroom-100.png',
+    'icons8-olive-oil-100.png',
+    'icons8-onion-100.png',
+    'icons8-orange-100.png',
+    'icons8-paprika-100.png',
+    'icons8-potato-100.png',
+    'icons8-prawn-100.png',
+    'icons8-pumpkin-100.png',
+    'icons8-salami-100.png',
+    'icons8-salami-pizza-100.png',
+    'icons8-salmon-100.png',
+    'icons8-strawberry-100.png',
+    'icons8-sugar-sack-100-2.png',
+    'icons8-sugar-sack-100.png',
+    'icons8-thanksgiving-turkey-100.png',
+    'icons8-tin-can-100.png',
+    'icons8-tomato-100.png',
+    'icons8-wine-and-grapes-100.png',
+  ].map((image) => {
+    const foodName = image.replace('icons8-', '').replace('-100.png', '');
+    return {
+      image,
+      name: foodName.replace('-', ''),
+    };
+  });
+}
